cleaned up the code & added missing types
